Add files via upload
This zip file contains the dataset , the jupyter notebook ( where the model is created ). I've used a linear regression model to predict the price of 22 essential commodities .
This project is divided into  3 parts --- i)  Cleaning and processing the data
                                                             ii)  EDA ( Exp. Data Analysis ) 
                                                             iii) Implementation of model.
Feel free to suggest me  on how to improve . Or suggest me  some more powerful efficient model for this project.